requirements.txt
# Unified requirements.txt for Multimodal RAG System

python-dotenv
pydantic
fastapi
uvicorn
requests
pillow
numpy
pdf2image
python-multipart
pymongo
streamlit
torch
sentence-transformers
fitz
pytesseract
pytest
black
isort
flake8
